fix: move speaker links to hole speaker dashboard
